Move status endpoint to dedicated controller
Public controller is huge and unweildy.

Also adds an API only base controller for our API only endpoints to use.

Currently API only and ApplicationController share almost all the same functionality, stored in the IncludeController module.

Ideally IncludeController will be broken up into smaller pieces in future refactors.
