+++ app/controllers/api_controller.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# Base controller for our API endpoints
+class ApiController < ActionController::API
+  include IncludeController
+end
