+++ spec/routing/status_routing_spec.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+describe StatusController, type: :routing do
+  it { expect(get('/status')).to route_to('status#index', format: 'json') }
+  it { expect(get('/status.json')).to route_to('status#index', format: 'json') }
+end
